requirements.txt
discord.py
jishaku
async_timeout
asyncpg
youtube_dl
Pillow
dateparser
bs4
psutil
humanize
git+https://github.com/Rapptz/discord-ext-menus
